Archive release artifacts in release Jenkinsfile
The artifacts are copied from the downstream branch builds and archived
on the release pipeline. This gives us a list of release candidates.

Since we are archiving artifacts from the release pipeline's workspace,
we should also clear the workspace when we're done, to guarantee we're
not archiving files from previous builds.